+++ screens/Home.js
@@ -1,0 +1,11 @@
+import React, {useState} from "react";
+import { View } from "react-native";
+import { Text } from "react-native-elements";
+
+export default function Home(){
+    return(
+        <View>
+            <Text h2>ENTROU</Text>
+        </View>
+    )
+}
